requirements.txt
boto3==1.34.148
botocore==1.34.148
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
copulas==0.11.0
ctgan==0.10.1
deepecho==0.6.0
dnspython==2.6.1
Faker==26.0.0
filelock==3.15.4
fsspec==2024.6.1
graphviz==0.20.3
idna==3.7
imbalanced-learn==0.12.3
imblearn==0.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
plotly==5.23.0
psutil==6.0.0
pymongo==4.8.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rdt==1.12.2
requests==2.32.3
s3transfer==0.10.2
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.0
sdmetrics==0.15.0
sdv==1.15.0
six==1.16.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
torch==2.4.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xgboost==2.1.1
